// apps/frontend-v2/src/lib/api/model/clinicFactsDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { ClinicTreatmentDto } from "./clinicTreatmentDto";
import type { ClinicDoctorDto } from "./clinicDoctorDto";
import type { ClinicOfferDto } from "./clinicOfferDto";

export interface ClinicFactsDto {
  treatments: ClinicTreatmentDto[];
  warranty: string;
  doctors: ClinicDoctorDto[];
  process: string;
  days: string;
  location: string;
  paymentMethods: string[];
  languages: string[];
  offers: ClinicOfferDto[];
}