// apps/frontend-v2/src/lib/api/model/clinicFactSheetDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lean Commerce API
 * The AI Sales Agent CRM API Documentation
 * OpenAPI spec version: 1.0
 */
import type { ClinicFactSheetDtoLatest } from "./clinicFactSheetDtoLatest";
import type { ClinicFactSheetDtoApproved } from "./clinicFactSheetDtoApproved";

export interface ClinicFactSheetDto {
  /** @nullable */
  latest: ClinicFactSheetDtoLatest;
  /** @nullable */
  approved: ClinicFactSheetDtoApproved;
}